initial support for markdown and manpage titles
We aim to provide support for markdown and manpage titles, by respecting
them and not merging them with other lines.

A markdown title is a line starting with # followed by the actual title.
A manpage title is a line made of uppercase letters only.

While the new code works, some tests are not  passing  because  the  new
code causes some bugs.

Thus, the added code is commented untill we figure out a way to  improve
it and make it pass the tests.